Keep sub-second precision in the visible-accounts cache key
BalanceSheet::AccountTotals#cache_key truncated the share timestamp with
.to_i, so a share added within the same second as an earlier one left the
cached account ids stale and the new account missing from the member's
view. Use .to_f, matching the net worth series builders.

+++ app/models/balance_sheet/account_totals.rb
@@ -66,7 +66,7 @@ class BalanceSheet::AccountTotals
 
     # Returns the cache key for storing visible account IDs, invalidated on data updates.
     def cache_key
-      shares_version = user ? AccountShare.where(user: user).maximum(:updated_at)&.to_i : nil
+      shares_version = user ? AccountShare.where(user: user).maximum(:updated_at)&.to_f : nil
       family.build_cache_key(
         [ "balance_sheet_account_ids", user&.id, shares_version ].compact.join("_"),
         invalidate_on_data_updates: true
